+++ src/hooks/useInventoryVersionHistory/useInventoryVersionHistory.js
@@ -18,6 +18,7 @@ import { useStripes } from '@folio/stripes/core';
 
 import { getChangedFieldsList } from './getChangedFieldsList';
 import { getActionLabel } from './getActionLabel';
+import { ACTIONS } from './constants';
 
 export const versionsFormatter = (usersMap, intl, canViewUser) => (diffArray) => {
   const anonymousUserLabel = intl.formatMessage({ id: 'ui-inventory.versionHistory.anonymousUser' });
@@ -39,8 +40,8 @@ export const versionsFormatter = (usersMap, intl, canViewUser) => (diffArray) =>
   };
 
   return diffArray
-    .filter(({ action }) => action !== 'CREATE')
-    .map(({ eventDate, eventTs, userId, eventId, diff }) => ({
+    .map(({ action, eventDate, eventTs, userId, eventId, diff }) => ({
+      isOriginal: action === ACTIONS.CREATE,
       eventDate: formatDateTime(eventDate, intl),
       source: getSourceLink(userId),
       userName: getUserName(userId) || anonymousUserLabel,